CustomChromeWindowBehavior is now BorderlessWindowBehavior
